22:23 24.03 Reworked deletion of rooms. Implemented basic methods with work of rooms, floors, users and bunnies. Redesigned user registration page

+++ src/main/java/com/studenthoteltest/demo/controller/FloorsController.java
@@ -2,12 +2,13 @@ package com.studenthoteltest.demo.controller;
 
 
 import com.studenthoteltest.demo.dao.model.Floors;
+import com.studenthoteltest.demo.dao.repository.FloorsRepository;
 import com.studenthoteltest.demo.service.FloorsService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
-import org.springframework.validation.BindingResult;
 import org.springframework.web.bind.annotation.*;
+import org.springframework.web.servlet.mvc.support.RedirectAttributes;
 
 import javax.validation.Valid;
 
@@ -15,6 +16,8 @@ import javax.validation.Valid;
 public class FloorsController {
     @Autowired
     private FloorsService floorsService;
+    @Autowired
+    private FloorsRepository floorsRepository;
 
     @GetMapping("/floors")
     public String floor(Model model){
@@ -26,7 +29,7 @@ public class FloorsController {
 
     @PostMapping("/floors")
     public String floorAdd(@Valid @ModelAttribute("floorAdd") Floors floor,
-                           BindingResult bindingResult,Model model
+                           Model model
     ){
         if (!floorsService.saveFloor(floor)){
             model.addAttribute("floorError","Такой этаж уже существует");
@@ -36,35 +39,29 @@ public class FloorsController {
         return "redirect:/floors";
     }
 
-    @GetMapping("/floors/gt/{floorsId}")
-    private String  getFloor(@PathVariable("floorId") Long florId, Model model) {
-        model.addAttribute("allFloors", floorsService.florList(florId));
-        return "floors";
-    }
+        @GetMapping("/floors/gt/{floorsId}")
+        private String  getFloor(@PathVariable("floorId") Long florId, Model model) {
+            model.addAttribute("allFloors", floorsService.florList(florId));
+            return "floors";
+        }
 
-//
-//
-
-
-    @GetMapping("/floorDelete")
-    public String floordelete(Model model){
-        model.addAttribute("allFloors", floorsService.allFloors());
-        return "floorDelete";
-    }
-
+//delete floor
+    //1
     @PostMapping("/floorDelete")
-    private String  deleteUser(@RequestParam(required = true, defaultValue = "" ) Long floorId,
-                               @RequestParam(required = true, defaultValue = "" ) String action) {
+    private String  deleteFloor(@RequestParam(required = true, defaultValue = "" ) Long floorId,
+                                @RequestParam(required = true, defaultValue = "" ) String action,
+                                Model model, RedirectAttributes redirectAttributes) {
+
         if (action.equals("delete")){
+            if (!floorsService.checkRoomsAtFloor(floorId)){
+                model.addAttribute("forAdmin","Вы не можите удалить этаж пока не разберетесь с камнатами, жильцами и заявками на заселение ");
+                redirectAttributes.addFlashAttribute("forAdmin","Вы не можите удалить этаж пока не разберетесь с камнатами, жильцами и заявками на заселение ");
+                return "redirect:/floors";
+            }
             floorsService.deleteFloor(floorId);
         }
-        return "redirect:/floorDelete";
+        return "redirect:/floors";
     }
-    @GetMapping("/floorDelete/gt/{floorsId}")
-    private String  gtFloorDelete(@PathVariable("floorId") Long florId, Model model) {
-        model.addAttribute("allFloors", floorsService.florList(florId));
-        return "floorDelete";
-    }
-
+    //
 
 }
